requirements.txt
appnope==0.1.0
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.2
cloudpickle==1.3.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
fasteners==0.15
flake8==3.8.3
future==0.18.2
glfw==1.12.0
gym==0.17.2
imageio==2.9.0
ipdb==0.13.3
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
kiwisolver==1.2.0
matplotlib==3.3.1
mccabe==0.6.1
monotonic==1.5
mujoco-py==2.0.2.13
numpy==1.19.5
parso==0.7.1
path==16.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pybullet==3.2.5
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pyglet==1.5.0
Pygments==2.6.1
pyparsing==2.4.7
pyquaternion==0.9.9
python-dateutil==2.8.1
scipy==1.5.2
six==1.15.0
torch==1.8.1
tqdm==4.48.2
traitlets==4.3.3
wcwidth==0.2.5
wandb
h5py
torchvision
opencv-python
colorlog
scikit-image
# tensorflow
